CSP-1308: Introduced a new integration test project.
