+++ src/DAS.AAN.Hub.Jobs.IntegrationTests/JobAudits/JobAuditWithRecentStartDate.cs
@@ -1,32 +1,23 @@
-﻿using Microsoft.EntityFrameworkCore;
-using SFA.DAS.AAN.Hub.Data;
+﻿using NUnit.Framework;
 using SFA.DAS.AAN.Hub.Data.Entities;
 using SFA.DAS.AAN.Hub.Data.Repositories;
+using SFA.DAS.AAN.Hub.Jobs.IntegrationTests.DbContext;
 
-namespace SFA.DAS.AAN.Hub.Jobs.UnitTests.Repositories
+namespace SFA.DAS.AAN.Hub.Jobs.IntegrationTests.JobAudits
 {
-    public class JobAuditRepositoryTests
+    public class JobAuditWithRecentStartDate
     {
-        private DbContextOptions<AanDataContext> _dbContextOptions;
         private CancellationToken cancellationToken = CancellationToken.None;
 
-        [SetUp]
-        public void Setup()
-        {
-            _dbContextOptions = new DbContextOptionsBuilder<AanDataContext>()
-                .UseInMemoryDatabase(databaseName: nameof(AanDataContext))
-                .Options;
-        }
-
         [Test]
-        public async Task ThenGetsMostRecentJobAudit()
+        public async Task Job_audit_with_todays_date_gets_returned()
         {
             JobAudit resultOne = new JobAudit() { JobName = nameof(JobAudit), StartTime = DateTime.UtcNow, EndTime = DateTime.UtcNow.AddMinutes(5) };
             JobAudit resultTwo = new JobAudit() { JobName = nameof(JobAudit), StartTime = DateTime.UtcNow.AddDays(-1), EndTime = DateTime.UtcNow.AddDays(-1).AddMinutes(5) };
 
             JobAudit? result;
 
-            using (var context = new AanDataContext(_dbContextOptions))
+            using (var context = InMemoryAanDataContext.CreateInMemoryContext($"{nameof(Job_audit_with_todays_date_gets_returned)}_InMemoryContext"))
             {
                 await context.JobAudits.AddRangeAsync([resultOne, resultTwo]);
                 await context.SaveChangesAsync(cancellationToken);
@@ -35,7 +26,7 @@ namespace SFA.DAS.AAN.Hub.Jobs.UnitTests.Repositories
             }
 
             Assert.That(result, Is.Not.Null);
-            Assert.That(result.StartTime.Date, Is.EqualTo(DateTime.UtcNow.Date));
+            Assert.That(result?.StartTime.Date, Is.EqualTo(DateTime.UtcNow.Date));
         }
     }
 }
