--- src/SFA.DAS.AAN.Hub.Jobs.UnitTests/Repositories/MemberRepositoryTests.cs
@@ -1,86 +1,0 @@
-﻿using AutoFixture;
-using Microsoft.EntityFrameworkCore;
-using SFA.DAS.AAN.Hub.Data;
-using SFA.DAS.AAN.Hub.Data.Entities;
-using SFA.DAS.AAN.Hub.Data.Repositories;
-
-namespace SFA.DAS.AAN.Hub.Jobs.UnitTests.Repositories;
-
-public class MemberRepositoryTests
-{
-    private CancellationToken cancellationToken = CancellationToken.None;
-    private Fixture _fixture;
-
-    [SetUp]
-    public void Setup()
-    {
-        _fixture = new Fixture();
-        _fixture.Behaviors.Remove(new ThrowingRecursionBehavior());
-        _fixture.Behaviors.Add(new OmitOnRecursionBehavior());
-    }
-
-    [Test]
-    public async Task ThenGetActiveApprenticeMembers()
-    {
-        var _dbContextOptions = new DbContextOptionsBuilder<AanDataContext>()
-            .UseInMemoryDatabase(databaseName: $"{nameof(AanDataContext)}_ThenGetActiveApprenticeMembers")
-            .Options;
-
-        var membersToAdd = _fixture.Build<Member>()
-                                        .Without(m => m.MemberPreferences)
-                                        .Without(m => m.Audits)
-                                        .Without(m => m.Notifications)
-                                        .Without(m => m.MemberProfiles)
-                                    .CreateMany(3)
-                                    .ToList();
-
-        List<Member> result;
-
-        using (var context = new AanDataContext(_dbContextOptions))
-        {
-            await context.AddRangeAsync(membersToAdd);
-            await context.SaveChangesAsync(cancellationToken);
-
-            var sut = new MemberRepository(context);
-            result = await sut.GetActiveApprenticeMembers(cancellationToken);
-        }
-
-        var apprenticeCount = membersToAdd.Where(a => a.UserType == UserType.Apprentice).Count();
-
-        Assert.That(result, Is.Not.Null);
-        Assert.That(result, Has.Count.EqualTo(apprenticeCount));
-    }
-
-    [Test]
-    public async Task WhenThereIsAnInactiveMember_ThenOnlyGetActiveMembers()
-    {
-        var _dbContextOptions = new DbContextOptionsBuilder<AanDataContext>()
-            .UseInMemoryDatabase(databaseName: $"{nameof(AanDataContext)}_ThenOnlyGetActiveMembers")
-            .Options;
-
-        var membersToAdd = _fixture.Build<Member>()
-                                        .With(m => m.UserType, UserType.Apprentice)
-                                        .Without(m => m.MemberPreferences)
-                                        .Without(m => m.Audits)
-                                        .Without(m => m.Notifications)
-                                        .Without(m => m.MemberProfiles)
-                                    .CreateMany(3)
-                                    .ToList();
-
-        membersToAdd[2].Email = membersToAdd[2].Id.ToString();
-
-        List<Member> result;
-
-        using (var context = new AanDataContext(_dbContextOptions))
-        {
-            await context.AddRangeAsync(membersToAdd);
-            await context.SaveChangesAsync(cancellationToken);
-
-            var sut = new MemberRepository(context);
-            result = await sut.GetActiveApprenticeMembers(cancellationToken);
-        }
-
-        Assert.That(result, Is.Not.Null);
-        Assert.That(result, Has.Count.EqualTo(2));
-    }
-}
