+++ src/SFA.DAS.AANHub.Jobs.sln
@@ -1,4 +1,3 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
@@ -10,6 +9,8 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{DCE3D45B-F3A0-42A1-89E1-88E2DAE64B68}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.AAN.Hub.Jobs.UnitTests", "SFA.DAS.AAN.Hub.Jobs.UnitTests\SFA.DAS.AAN.Hub.Jobs.UnitTests.csproj", "{BC1D52F4-6180-40BA-BF1B-5547E7FB2E3A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SFA.DAS.AAN.Hub.Jobs.IntegrationTests", "DAS.AAN.Hub.Jobs.IntegrationTests\SFA.DAS.AAN.Hub.Jobs.IntegrationTests.csproj", "{DFE656B9-B59B-440A-8AAA-49651B811B8A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -29,12 +30,17 @@ Global
 		{BC1D52F4-6180-40BA-BF1B-5547E7FB2E3A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BC1D52F4-6180-40BA-BF1B-5547E7FB2E3A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BC1D52F4-6180-40BA-BF1B-5547E7FB2E3A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DFE656B9-B59B-440A-8AAA-49651B811B8A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DFE656B9-B59B-440A-8AAA-49651B811B8A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DFE656B9-B59B-440A-8AAA-49651B811B8A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DFE656B9-B59B-440A-8AAA-49651B811B8A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{BC1D52F4-6180-40BA-BF1B-5547E7FB2E3A} = {DCE3D45B-F3A0-42A1-89E1-88E2DAE64B68}
+		{DFE656B9-B59B-440A-8AAA-49651B811B8A} = {DCE3D45B-F3A0-42A1-89E1-88E2DAE64B68}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {44DEF13B-4DFC-4EF1-A33E-8F090697EBCC}
